requirements.txt
python-dotenv
pandas
matplotlib
requests